moves data fetching function and related code out of component file
